requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
beautifulsoup4
python-dotenv
faker
geopy
selenium
webdriver_manager
apscheduler